Added a toggle to turn of things doing damage to the player
Unfortunately,  I haven't figured out how to get static variables to be visible in the inspector, so if you want to use this feature, you must manually change the attackingOn variable in the start function. Note: a side effect of this is that you are able to destroy spike traps, which is not typically seen because the player dies before this happens.  It is fine though.